refactor(watch): wrap iPhone WCSession bridge in config plugin
