+++ apps/mobile/.gitignore
@@ -18,12 +18,10 @@ GoogleService-Info.plist
 # Google Play Service Account
 play-store-service-account.json
 
-# iOS (generated by expo prebuild — only track specific hand-edited files)
+# iOS (generated by expo prebuild — only track Info.plist which holds version/build metadata)
 ios/**
 !ios/ProstCounter/
 !ios/ProstCounter/Info.plist
-!ios/ProstCounter/AppDelegate.swift
-!ios/ProstCounter/WatchSessionBridge.swift
 
 # Android
 android/
