+++ apps/mobile/app.config.ts
@@ -138,6 +138,7 @@ export default ({ config }: ConfigContext): ExpoConfig => ({
     ],
     "./plugins/withModularHeaders.js",
     "./plugins/withApsEnvironment.js",
+    "./plugins/withWatchSessionBridge.js",
     "./plugins/withFirebaseNotificationColor.js",
     "./plugins/withGoogleMapsApiKey.js",
     "./plugins/withAsyncStorageMavenRepo.js",
