+++ apps/mobile/plugins/withWatchSessionBridge.js
@@ -1,0 +1,164 @@
+const { withDangerousMod } = require("@expo/config-plugins");
+const fs = require("fs");
+const path = require("path");
+
+/**
+ * Config plugin that injects the WatchSessionBridge class into AppDelegate.swift
+ * and activates it in application(_:didFinishLaunchingWithOptions:).
+ *
+ * Why merged into AppDelegate.swift instead of a separate file:
+ *   AppDelegate.swift is already in the ProstCounter target's Sources build phase.
+ *   Merging here avoids patching project.pbxproj, which is brittle. The class
+ *   survives `expo prebuild --clean` automatically since it is regenerated by this plugin.
+ */
+
+const BRIDGE_CLASS_SNIPPET = `
+// MARK: - WatchSessionBridge (injected by withWatchSessionBridge.js)
+
+import WatchConnectivity
+
+/// Bridges App Group UserDefaults changes on the iPhone to the paired Apple Watch
+/// via WCSession.transferUserInfo.
+///
+/// Flow:
+///   JS (ExtensionStorage) → UserDefaults(suiteName:) → KVO here → WCSession.transferUserInfo
+final class WatchSessionBridge: NSObject {
+
+  static let shared = WatchSessionBridge()
+
+  private let appGroup = "group.com.prostcounter.shared"
+  private let watchedKeys = ["accessToken", "refreshToken", "userId", "currentFestivalId", "expiresAt"]
+  private var defaults: UserDefaults?
+
+  private override init() {
+    super.init()
+    guard WCSession.isSupported() else { return }
+    defaults = UserDefaults(suiteName: appGroup)
+    WCSession.default.delegate = self
+    WCSession.default.activate()
+    startObserving()
+  }
+
+  private func startObserving() {
+    guard let defaults = defaults else { return }
+    NotificationCenter.default.addObserver(
+      self,
+      selector: #selector(defaultsDidChange(_:)),
+      name: UserDefaults.didChangeNotification,
+      object: defaults
+    )
+  }
+
+  @objc private func defaultsDidChange(_ notification: Notification) {
+    forwardToWatch()
+  }
+
+  private func forwardToWatch() {
+    guard
+      WCSession.default.activationState == .activated,
+      WCSession.default.isPaired,
+      WCSession.default.isWatchAppInstalled,
+      let defaults = defaults
+    else { return }
+
+    var payload: [String: Any] = [:]
+    for key in watchedKeys {
+      if let value = defaults.object(forKey: key) {
+        payload[key] = value
+      }
+    }
+    guard !payload.isEmpty else { return }
+
+    WCSession.default.transferUserInfo(payload)
+  }
+
+  deinit {
+    NotificationCenter.default.removeObserver(self)
+  }
+}
+
+// MARK: - WCSessionDelegate
+
+extension WatchSessionBridge: WCSessionDelegate {
+
+  func session(
+    _ session: WCSession,
+    activationDidCompleteWith activationState: WCSessionActivationState,
+    error: Error?
+  ) {
+    if let error = error {
+      print("[WatchSessionBridge] Activation error: \\(error)")
+      return
+    }
+    if activationState == .activated {
+      forwardToWatch()
+    }
+  }
+
+  func sessionDidBecomeInactive(_ session: WCSession) {
+    // Required on iOS — nothing to do
+  }
+
+  func sessionDidDeactivate(_ session: WCSession) {
+    // Required on iOS — reactivate to support Apple Watch switching
+    WCSession.default.activate()
+  }
+}
+`;
+
+// Inserted right after the #endif that closes the #if os(iOS) || os(tvOS) block
+// in application(_:didFinishLaunchingWithOptions:).
+// Anchor: the literal "#endif" line that follows factory.startReactNative(...)
+const INIT_CALL =
+  "\n    // Activate the WatchConnectivity bridge so session tokens written to the\n    // shared App Group are forwarded to the paired Apple Watch.\n    _ = WatchSessionBridge.shared\n";
+
+/**
+ * Anchor used for the init-call insertion.
+ * In the Expo-generated AppDelegate.swift the #if os(iOS) || os(tvOS) block ends with
+ * exactly this string. We append the activation call immediately after it.
+ */
+const INIT_ANCHOR = "#endif";
+
+module.exports = function withWatchSessionBridge(config) {
+  return withDangerousMod(config, [
+    "ios",
+    async (cfg) => {
+      const appDelegatePath = path.join(
+        cfg.modRequest.platformProjectRoot,
+        "ProstCounter",
+        "AppDelegate.swift",
+      );
+
+      let src = fs.readFileSync(appDelegatePath, "utf-8");
+
+      if (src.includes("WatchSessionBridge")) {
+        console.log(
+          "withWatchSessionBridge: AppDelegate.swift already contains WatchSessionBridge — skipping.",
+        );
+        return cfg;
+      }
+
+      // 1. Inject the activation call after the closing #endif of the
+      //    #if os(iOS) || os(tvOS) block inside didFinishLaunchingWithOptions.
+      //    The generated file has exactly one "#endif" inside that function.
+      //    We replace the first occurrence to be safe and precise.
+      if (!src.includes(INIT_ANCHOR)) {
+        console.warn(
+          "withWatchSessionBridge: Could not find '#endif' anchor in AppDelegate.swift — init call NOT injected.",
+        );
+      } else {
+        src = src.replace(INIT_ANCHOR, INIT_ANCHOR + INIT_CALL);
+      }
+
+      // 2. Append the WatchSessionBridge class as top-level code at end of file.
+      src = src + "\n" + BRIDGE_CLASS_SNIPPET;
+
+      fs.writeFileSync(appDelegatePath, src, "utf-8");
+      console.log(
+        "withWatchSessionBridge: Injected WatchSessionBridge into AppDelegate.swift.",
+      );
+
+      return cfg;
+    },
+  ]);
+};
